// src/app/models/ecoles.ts
export interface Ecoles {

  id:number;
  nomE: string;
  groupeScolaire: string;
  adressePostale: string;
  pays:string;
  tel1: string;
  tel2: string;
  email1:string;
  email2: string;
  siteWeb: string;
  presentation:string;
  titrePresentation:string;
  chefEtablissement:string;
  photoEtablissement: string;
  signiature: string;
  creePar:number;
  modifiePar:number;
  createdAt: string;
  updatedAt: string;
  idGroupeScolaire:number;
  imgPageCnx: string;
  logoPageCnx: string;
  couleurExterne: string;
  logoBarreTache: string;
  couleurBarreTache: string;
  nbrEleve:string;
  nbrParent:string;
  nbrGroupe:string;
  nbrEcole:string;
  nomG:string;
}

/*
 export interface Ecoles {
 idE:number;
 nomE: string;
 adressePostale: string;
 pays:string;
 tel1: string;
 tel2: string;
 email1:string;
 email2: string;
 siteWeb: string;
 chefEtablissement:string;
 photoEtablissement: string;
 signiature: string;
 presentation:string;
 creePar:number;
 modifiePar:number;
 createdAt: string;
 updatedAt: string;
 idGroupeScolaire:number;
 }
 */